Allow more space for the Greek equivalent of "of", as submitted by Yianni:
I'm using the Greek translation and in the Videos-->Videos List screen
(classic layout) the "x of y" textarea is slightly less wide than necessary,
because the Greek word for "of" is 3 letters long. I've taken the liberty to
expand the position width and decrease the breadcrumbs one by 20 units.