OSD NAV: Start with focus on Pause.
This is because that is the most often urgently needed
action, for example when watching s video and the doorbell rings,
somebody starts talking to you or the cat vomits on the carpet. You
don't want to be pressing left and right looking for Pause when your
attention is elsewhere. Better to press the enter button twice and you
are paused.

Signed-off-by: John Poet <jpoet@mythtv.org>
(cherry picked from commit 700da9f715305078b1acf7deca639f94f8bcc5f2)